start client session handler
